implement env file and restructure project
