+++ .env.example
@@ -1,20 +1,27 @@
 # Your Gmail for creating calendar events
 MY_CALENDAR_EMAIL='me@gmail.com'
 
-# Path to credentials directory
-# where credentials.json is stored
+# Path to credentials directory where credentials.json is stored
 CREDENTIALS_DIR=./credentials
-
-# Google doc and sheets URL ids for testing
-GDOC_ID_DEV='your-test-gdoc-id'
-GSHEET_ID_DEV='your-test-gsheet-id'
 
 # Google doc and sheets URL ids for production
 GDOC_ID_PROD='your-prod-gdoc-id'
 GSHEET_ID_PROD='your-prod-gsheet-id'
 
-# Dinner attendees that will receiver calendar events, etc.
-DINNER_ATTENDEES={
-    'Attendee 1 name':'attendee1@gmail.com',
-    'Attendee 2 name':'attendee2@gmail.com'
-}
+# Google doc and sheets URL ids for testing
+GDOC_ID_DEV='your-test-gdoc-id'
+GSHEET_ID_DEV='your-test-gsheet-id'
+
+# stringified dinner attendees dictionary that will receive calendar events, etc.
+#{
+#    'Attendee 1 name':'attendee1@gmail.com',
+#    'Attendee 2 name':'attendee2@gmail.com'
+#}
+DINNER_ATTENDEES='{"Attendee 1 name": "attendee1@gmail.com", "Attendee 2 name": "attendee2@gmail.com"}'
+
+# stringified dictionary mapping notify_who text from gsheet to [notification type], [how many mins before event to notify]
+#{
+#    'x hours before as popup': ['popup', 'x*60'],
+#    'at time of event as email': ['email', '0']
+#}
+REMINDERS='{"x hours before as popup": ['popup', 'x*60'], "at time of event as email": ['email', '0']}'
