+++ helpers/color_codes.py
@@ -1,10 +1,10 @@
 # color codes for text
-COLOR_RED = [0.0,0.0,1.0] # first week
-COLOR_PURPLE = [0.9,0.1,0.6] # second week
-COLOR_YELLOW = [0.0,1.0,1.0] # test
-COLOR_GREEN = [0.0,1.0,0.0]
-COLOR_BLUE = [1.0,0.0,0.0]
-COLOR_BLACK = [0.0,0.0,0.0]
+FONT_RED = [0.0,0.0,1.0] # first week
+FONT_PURPLE = [0.9,0.1,0.6] # second week
+FONT_YELLOW = [0.0,1.0,1.0] # test
+FONT_GREEN = [0.0,1.0,0.0]
+FONT_BLUE = [1.0,0.0,0.0]
+FONT_BLACK = [0.0,0.0,0.0]
 
 # color codes for calendar events
 EVENT_LAVENDER = '1'
